+++ Changes.md
@@ -6,6 +6,11 @@ Fixes
 
 - Windows `Scene/OpenGL/Shader` Menu : Removed `\` at the beginning of menu items.
 
+API
+---
+
+- SceneGadget : Added `snapshotToFile()` method.
+
 1.3.7.0 (relative to 1.3.6.1)
 =======